[REFACTOR] 마이페이지 카테고리 파트 리팩토링
